Move non primary modules into utils folder.
