+++ spec/src/modules/tracker-humanity.js
@@ -1,8 +1,8 @@
 const dotenv = require('dotenv');
 const chai = require('chai');
 const chaiAsPromised = require('chai-as-promised');
-const store = require('../../../src/store/store');
-const trackerHumanity = require('../../../src/modules/tracker-humanity');
+const store = require('../../../src/utils/store');
+const trackerHumanity = require('../../../src/utils/tracker-humanity');
 const helpers = require('../../mocha.helpers');
 
 chai.use(chaiAsPromised);
